Fixed some warnings, and optimized the hairs file writing. Now only 1 hairs.rib is written, instead of many separate files.
git-svn-id: http://makehuman.googlecode.com/svn@182 ea038faa-237f-11de-8bc4-bb72633a7edc

+++ trunk/makehuman/mh_plugins/hairgenerator.py
@@ -368,9 +368,9 @@ class Hairgenerator:
             elif datalist[0] == "tipMagnet":
                 self.tipMagnet = float(datalist[1])
             elif datalist[0] == "numberOfHairsClump":
-                self.numberOfHairsClump = float(datalist[1])
+                self.numberOfHairsClump = int(datalist[1])
             elif datalist[0] == "numberOfHairsMultiStrand":
-                self.numberOfHairsMultiStrand = float(datalist[1])
+                self.numberOfHairsMultiStrand = int(datalist[1])
             elif datalist[0] == "randomFactClump":
                 self.randomFactClump = float(datalist[1])
             elif datalist[0] == "randomFactMultiStrand":
